🛠️ experiments/gcg_exp.py -> renamed to poisoncraft_exp.py 🛠️ src/GCG/__init__.py -> renamed to poisoncraft/__init__.py 🛠️ src/GCG/gcg.py -> renamed to poisoncraft/gcg.py 🛠️ src/GCG/gcg_utils.py -> renamed to poisoncraft/gcg_utils.py 🛠️ src/GCG/run_gcg.py -> renamed to poisoncraft/run_gcg.py
