+++ experiments/poisoncraft_exp.py
@@ -2,9 +2,9 @@ import sys
 import os
 import pdb
 
-sys.path.append(os.path.abspath('../RAGInjection/'))
+sys.path.append(os.path.abspath('../PoisonCraft/'))
 
-from src.GCG.run_gcg import gcg_attack
+from src.poisoncraft.run_gcg import gcg_attack
 from src.utils import load_frequent_words_as_str
 
 import argparse
